# Conflicts:
#	README.md
#	bins/revme/Cargo.toml
#	bins/revme/src/cmd/statetest/runner.rs
#	crates/interpreter/Cargo.toml
#	crates/interpreter/src/instructions/contract.rs
#	crates/interpreter/src/interpreter.rs
#	crates/precompile/Cargo.toml
#	crates/primitives/Cargo.toml
#	crates/primitives/src/env.rs
#	crates/primitives/src/lib.rs
#	crates/primitives/src/result.rs
#	crates/primitives/src/state.rs
#	crates/revm/Cargo.toml
#	crates/revm/src/builder.rs
#	crates/revm/src/context/evm_context.rs
#	crates/revm/src/context/inner_evm_context.rs
#	crates/revm/src/db.rs
#	crates/revm/src/evm.rs
#	crates/revm/src/handler/mainnet/post_execution.rs
#	crates/revm/src/inspector.rs
#	crates/revm/src/journaled_state.rs
#	crates/revm/src/optimism/fast_lz.rs
#	e2e/Cargo.lock
#	examples/fork_ref_transact.rs
#	examples/generate_block_traces.rs

+++ crates/primitives/src/constants.rs
@@ -12,7 +12,7 @@ pub const MAX_CODE_SIZE: usize = 0x6000;
 pub const WASM_MAX_CODE_SIZE: usize = 0x100000;
 
 /// Number of block hashes that EVM can access in the past (pre-Prague).
-pub const BLOCK_HASH_HISTORY: usize = 256;
+pub const BLOCK_HASH_HISTORY: u64 = 256;
 
 /// EIP-2935: Serve historical block hashes from state
 ///
